HACKEVEN/NHCEPS-3/front/requirements.txt
flask==2.0.1
opencv-python==4.5.3.56
pytesseract==0.3.8
python-dotenv==0.19.0
openai==0.27.0
aiohttp==3.8.1
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0 